feat: update redirection notification for external links
